services/catalog-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy>=2.0
alembic
pydantic==2.9.2
pydantic-settings==2.6.1
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
pytest
httpx
pytest-cov
